% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_genetic_positions.R
\name{interpolate_genetic_position}
\alias{interpolate_genetic_position}
\title{Predict marker genetic position using linear interpolation}
\usage{
interpolate_genetic_position(geno, cMMb)
}
\arguments{
\item{geno}{A data frame of marker allele dosages. See \italic{Details} for more information.}

\item{cMMb}{A data frame of the recombination rate information. See \italic{Details}.}
}
\description{
Uses linear interpolation and local recombination rate information to predict
the unknown genetic position of markers based on their known physical positions.
}
\details{
The first columns of the geno object are marker, chrom, and bp. Subsequent columns
contain the allele dosage for individuals/clones, coded 0,1,2. While fractional values are allowed for genomewide prediction,
these values are rounded when predicting the genetic variance within crosses.

The cMMb object must contains six columns: chrom, left_bp (the physical position of the left
flank of an interval), right_bp (the physical position of the right flank of an interval),
left_cM (the genetic position of the left flank of an interval), right_cM
(the genetic position of the right flank of an interval), and cM_Mb (the recombination rate
(in cM / Mb) within that interval).
}
