% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_predict2.R
\name{pop.predict2}
\alias{pop.predict2}
\alias{pop_predict2}
\title{Predict genetic variance and genetic correlations in bi-parental populations using a deterministic model}
\usage{
pop.predict2(G.in, y.in, map.in, crossing.table, parents, tail.p = 0.1,
  self.gen = Inf, DH = FALSE, model = c("rrBLUP", "BayesC"))

pop_predict2(M, y.in, marker.effects, map.in, crossing.table, parents,
  tail.p = 0.1, self.gen = Inf, DH = FALSE, model = c("rrBLUP",
  "BayesC"))
}
\arguments{
\item{G.in}{See \code{G.in} in \code{\link[PopVar]{pop.predict}}.}

\item{y.in}{See \code{y.in} in \code{\link[PopVar]{pop.predict}}.}

\item{map.in}{See \code{map.in} in \code{\link[PopVar]{pop.predict}}.}

\item{crossing.table}{See \code{crossing.table} in \code{\link[PopVar]{pop.predict}}.}

\item{parents}{See \code{parents} in \code{\link[PopVar]{pop.predict}}.}

\item{tail.p}{See \code{tail.p} in \code{\link[PopVar]{pop.predict}}.}

\item{self.gen}{The number of selfing generations in the potential cross. Can be an integer or \code{Inf} for
recombinant inbreds. Note: \code{self.gen = 1} corresponds to an F2 population.}

\item{DH}{Indicator if doubled-haploids are to be induced after the number of selfing generations indicated by
\code{self.gen}. For example, if \code{self.gen = 0} and \code{DH = TRUE}, then doubled-haploids are asssumed
to be induced using gametes from F1 plants.}

\item{model}{See \code{models} in \code{\link[PopVar]{pop.predict}}. Only 1 model is allowed.}

\item{M}{A Matrix of marker genotypes of dimensions \code{nLine} x \code{nMarker}, coded as
-1, 0, and 1.}

\item{marker.effects}{A data frame of marker effects. The first column should include the marker name and
subsequent columns should include the marker effects. Supercedes \code{y.in} if passed.}
}
\description{

}
\details{
Predictions are based on the deterministic equations specified by Zhong and Jannink (2007), Allier et al. (2019),
and Neyhart et al. (2019).
}
\section{Functions}{
\itemize{
\item \code{pop_predict2}: 
}}

\examples{

}
\references{
Zhong, S., and J.-L. Jannink, 2007 Using quantitative trait loci results to discriminate among crosses on the basis of their
progeny mean and variance. Genetics 177: 567–576. https://doi.org/10.1534/ genetics.107.075358

Allier, A., L. Moreau, A. Charcosset, S. Teyssèdre, and C. Lehermeier, 2019 Usefulness Criterion and Post-selection Parental
Contributions in Multi-parental Crosses: Application to Polygenic Trait Introgression. G3 9: 1469–1479.
doi: 10.1534/g3.119.400129

Neyhart, J.L., A.J. Lorenz, and K.P. Smith, 2019 Multi-trait Improvement by Predicting Genetic Correlations in Breeding
Crosses. G3 9: 3153-3165. doi: 10.1534/g3.119.400406
}
