% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopVar-package.R
\docType{package}
\name{PopVar-package}
\alias{PopVar}
\alias{PopVar-package}
\title{PopVar: Genomic Breeding Tools: Genetic Variance Prediction and Cross-Validation}
\description{
The main attribute of 'PopVar' is the prediction of genetic variance in bi-parental populations, from which the package derives its name. 'PopVar' contains a set of functions that use phenotypic and genotypic data from a set of candidate parents to 1) predict the mean, genetic variance, and superior progeny value of all, or a defined set of pairwise bi-parental crosses, and 2) perform cross-validation to estimate genome-wide prediction accuracy of multiple statistical models. More details are available in Mohammadi, Tiede, and Smith (2015, \doi{10.2135/cropsci2015.01.0030}). A dataset 'think_barley.rda' is included for reference and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UMN-BarleyOatSilphium/PopVar}
  \item Report bugs at \url{https://github.com/UMN-BarleyOatSilphium/PopVar/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey Neyhart \email{neyhartje@gmail.com} (\href{https://orcid.org/0000-0002-1991-5310}{ORCID}) [contributor]

Authors:
\itemize{
  \item Tyler Tiede \email{tyler.tiede7@gmail.com}
}

Other contributors:
\itemize{
  \item Mohsen Mohammadi (\href{https://orcid.org/0000-0002-4536-1200}{ORCID}) [contributor]
  \item Kevin Smith (\href{https://orcid.org/0000-0001-8253-3326}{ORCID}) [contributor]
}

}
\keyword{internal}
