% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PopVar.data-class}
\alias{PopVar.data-class}
\alias{PopVar.data}
\title{S4 class of genotypic and phenotypic data}
\description{
S4 class of genotypic and phenotypic data
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes coded as allele dosages z = { 0, 1, 2 }}

\item{\code{allele.freq}}{A numeric vector of allele frequencies of the first allele.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}

\item{\code{missing}}{Whether there is any missing marker data.}

\item{\code{assume.inbred}}{Whether to assume unphased marker data comes from completely homozygous individuals.}

\item{\code{dominance}}{Whether to compute dominance marker matrices and effects.}

\item{\code{pheno}}{A data frame of phenotypic data}

\item{\code{fixed}}{A data frame of fixed effects}

\item{\code{traits}}{A list of trait names}
}}

