% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PopVar.me-class}
\alias{PopVar.me-class}
\alias{PopVar.me}
\title{S4 class of genotypic and phenotypic data with estimated marker effects}
\description{
S4 class of genotypic and phenotypic data with estimated marker effects
}
\section{Slots}{

\describe{
\item{\code{marker.effects}}{A matrix of marker effects}
}}

