% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_marker_effects.R
\name{predict_marker_effects}
\alias{predict_marker_effects}
\title{Predict marker effects}
\usage{
predict_marker_effects(
  data,
  models = c("rrBLUP", "BayesA", "BayesB", "BayesC", "BL", "BRR"),
  params = list(nIter = 500, burnIn = 50)
)
}
\arguments{
\item{data}{A \code{PopVar.data} object from \code{read_PopVar}.}

\item{models}{The list of models to run. See Details for supported models.}

\item{params}{Parameters to pass to \code{\link{BGLR}}, if \code{models} includes Bayesian models.}
}
\value{
A variable of class \code{\link[PopVar.me]{PopVar.me}}
}
\description{
Predict marker effects
}
\details{
The function currently supports the models:
\describe{
\item{rrBLUP}{Ridge regression best linear unbiased prediction.}
\item{BayesA}{BayesA.}
\item{BayesB}{BayesB.}
\item{BayesC}{BayesC-pi.}
\item{BL}{Bayesian Lasso.}
\item{BRR}{Bayesian ridge regression.}
}
}
\examples{
# Read in unphased cranberry marker data and phenotypes
geno <- cranberry_geno
pheno <- cranberry_pheno
dat <- read_PopVar(geno = geno, pheno = pheno, n.traits = 6, phased = FALSE, 
                   assume.inbred = FALSE)

# Compute marker effects using the rrBLUP and RKHS models
data_me <- predict_marker_effects(data = popvar_data, models = c("rrBLUP", "BRR"))

}
