% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_merit.R
\name{predict_merit}
\alias{predict_merit}
\title{Predict breeding values or genotypic values}
\usage{
predict_merit(data)
}
\arguments{
\item{data}{A \code{PopVar.data} object from \code{read_PopVar}.}

\item{models}{The list of models to run. Currently supports ...}

\item{method}{The cross-validation method. Either fractional (\code{"frac"}), k-fold (\code{"kfold"}), or leave-one-out \code{"loo"}.}

\item{p.test}{For fractional cross-validation, the proportion of individuals to designate as the test set per replicate. \code{1 - p.test} is the proportion of individuals used for training.}

\item{k}{For k-fold cross-validation, the number of folds.}

\item{reps}{The number of complete replicates of fractional or k-fold cross-validation.}

\item{group}{The name of a column in the phenotype data for grouped cross-validation (e.g. environment, location, etc.). For example to run leave-one-environment-out cross-validation, set \code{method = "loo"} and \code{group = "env"}.}
}
\description{
Predict breeding values or genotypic values
}
