% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pop.R
\name{predict_pop}
\alias{predict_pop}
\title{Predict genetic variance and genetic correlations in bi-parental populations deterministic equations}
\usage{
predict_pop(
  data,
  crossing.table,
  parents,
  tail.p = 0.1,
  self.gen = 0,
  DH = FALSE
)
}
\arguments{
\item{data}{A \code{PopVar.me} object from \code{\link{predict_marker_effects}}.}

\item{crossing.table}{A data frame of specific crosses to predict. Each row is a cross and the first two columns are the parents of the cross.}

\item{parents}{A character vector of potential parents. If passed, \code{crossing.table} is ignored and all pairwise crosses using \code{parents} will be predicted.}

\item{tail.p}{The percentile of the simulated progeny to be included into the calculation of \eqn{\mu}\emph{_sp} and correlated response. Default is \code{0.10}.}

\item{self.gen}{The number of selfing generations in the potential cross. Can be an integer or \code{Inf} for
recombinant inbreds. Note: \code{self.gen = 0} corresponds to an F1 population for outbreds and \code{self.gen = 1} corresponds to an F2 population.}

\item{DH}{Indicator if doubled-haploids are to be induced after the number of selfing generations indicated by
\code{self.gen}. For example, if \code{self.gen = 0} and \code{DH = TRUE}, then doubled-haploids are assumed
to be induced using gametes from F1 plants.}
}
\description{
Predict genetic variance and genetic correlations in bi-parental populations deterministic equations
}
\details{
Predictions are based on the deterministic equations specified by Zhong and Jannink (2007), Allier et al. (2019),
and Neyhart et al. (2019).
}
\examples{

\donttest{

# Read in phased cranberry marker data and phenotypes
geno <- cranberry_geno_phased
pheno <- cranberry_pheno
dat <- read_PopVar(geno = geno, pheno = pheno, n.traits = 6, phased = TRUE, 
                   assume.inbred = FALSE, dominance = TRUE, sep = "_")

# Compute marker effects using the rrBLUP model
data_me <- predict_marker_effects(data = dat, models = c("rrBLUP"))

# Simulate a crossing table
parents <- colnames(slot(data_me, "geno.mat"))
parent1 <- sample(parents, 10)
parent2 <- sample(setdiff(parents, parent1), 10)
cross_tab <- data.frame(parent1 = parent1, parent2 = parent2)

# Use example data to make predictions
out <- pop.predict2(G.in = G.in_ex_imputed, y.in = y.in_ex, map.in = map.in_ex, 
                    crossing.table = cross.tab_ex)
                    
# Provide a vector of parents to predict all possible crosses (some parents
# have missing phenotypic data)
out <- pop.predict2(G.in = G.in_ex_imputed, y.in = y.in_ex, map.in = map.in_ex, 
                    parents = y.in_ex$Entry[1:5])
                    
# Make predictions for 5 crosses with various levels of inbreeding
out_list <- lapply(X = 1:10, FUN = function(self.gen) {
  out <- pop.predict2(G.in = G.in_ex_imputed, y.in = y.in_ex, map.in = map.in_ex, 
                      crossing.table = cross.tab_ex[1:5,], self.gen = self.gen)
  out$self.gen <- self.gen
  out })
               
# Plot predictions of grain yield genetic variance over levels of inbreeding
dat <- do.call("rbind", lapply(out_list, subset, trait == "Yield"))
plot(pred_varG ~ self.gen, data = dat, type = "b", 
     subset = parent1 == parent1[1] & parent2 == parent2[1])

}

}
\references{
Zhong, S., and J.-L. Jannink, 2007 Using quantitative trait loci results to discriminate among crosses on the basis of their
progeny mean and variance. Genetics 177: 567–576. https://doi.org/10.1534/ genetics.107.075358

Allier, A., L. Moreau, A. Charcosset, S. Teyssèdre, and C. Lehermeier, 2019 Usefulness Criterion and Post-selection Parental
Contributions in Multi-parental Crosses: Application to Polygenic Trait Introgression. G3 9: 1469–1479.
doi: 10.1534/g3.119.400129

Neyhart, J.L., A.J. Lorenz, and K.P. Smith, 2019 Multi-trait Improvement by Predicting Genetic Correlations in Breeding
Crosses. G3 9: 3153-3165. doi: 10.1534/g3.119.400406
}
