% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{read_geno}
\alias{read_geno}
\title{Read in marker genotype data}
\usage{
read_geno(
  filename,
  geno,
  ploidy = 2L,
  phased = FALSE,
  sep = ".",
  min.mac = 5,
  max.r = 1,
  dominance = FALSE,
  inbred = FALSE
)
}
\arguments{
\item{filename}{Name of CSV file with marker allele dosage}

\item{geno}{A data frame of marker allele dosages. If passed, \code{filename} is ignored.}

\item{ploidy}{The ploidy level e.g. 2, 4, 6, ..., etc.}

\item{phased}{If TRUE, the input genotype matrix is phased, with alleles coded as positive integers. The name of each
individual/clone column is the id and haplotype, concatenated by \code{sep}.}

\item{sep}{The haplotype concatenator; see \code{phased}.}

\item{min.mac}{The minimum minor allele count to retain a marker.}

\item{max.r}{The maximum correlation between any pair of markers. The \code{r^2} is used directly in filtering. For each pair of markers exceeding this threshold, the one with the greater minor allele frequency is retained. Set \code{max.r = 1} for no filtering.}

\item{dominance}{Logical. Should a dominance relationship matrix be constructed?}

\item{inbred}{Logical. If unphased marker data is passed, should the function assume that all genotypes are inbred?}
}
\value{
Variable of class \code{\link{PopVar.geno}}.
}
\description{
Read in marker genotype data
}
\details{
The first columns of the file or geno object are marker, chrom, cM (and optionally bp, if provided).Subsequent columns
contain the allele dosage for individuals/clones, coded 0,1,2. While fractional values are allowed for genomewide prediction,
these values are rounded when predicting the genetic variance within crosses.

Only diploid allele dosages are allowed.

When \code{dominance=FALSE}, non-additive effects are not captured.
If \code{dominance=TRUE}, a (digenic) dominance covariance matrix is used instead.
}
\examples{
# Read in barley marker genotypes assuming no phasing and non-inbred lines.
geno_in <- read_geno(geno = barley_geno, min.mac = 5)

# Read in barley marker genotypes assuming completely inbred lines
geno_in <- read_geno(geno = barley_geno, min.mac = 5, inbred = TRUE)

# Read in cranberry marker genotype assuming no phasing
geno_in <- read_geno(geno = cranberry_geno, min.mac = 5)

# Read in cranberry phased haplotypes from Beagle
geno_in <- read_geno(geno = cranberry_geno_phased, phased = TRUE, sep = "_hap", min.mac = 5)

}
\references{
Endelman, J.B. 2023. Fully efficient, two-stage analysis of multi-environment trials with directional dominance and multi-trait genomic selection. Theor. Appl. Genet. 136(4): 65.
}
