% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PopVar.data-class}
\alias{PopVar.data-class}
\alias{PopVar.data}
\title{S4 class of genotypic and phenotypic data}
\description{
S4 class of genotypic and phenotypic data
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}

\item{\code{missing}}{Whether there is any missing marker data.}

\item{\code{assume.inbred}}{Whether to assume unphased marker data comes from completely homozygous individuals.}

\item{\code{geno}}{A \code{\link{PopVar.geno}} object.}

\item{\code{pheno}}{A data frame of phenotypic data}

\item{\code{fixed}}{A data frame of fixed effects}

\item{\code{traits}}{A list of trait names}
}}

