% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_PopVar.R
\name{read_PopVar}
\alias{read_PopVar}
\title{Prepare phenotype and marker genotype data}
\usage{
read_PopVar(geno, pheno, n.traits, delim = ",")
}
\arguments{
\item{geno}{An object of class \code{PopVar.geno} from the \code{\link{read_geno}} function.}

\item{pheno}{A data frame or filename of the phenotypic data. The first column of the data frame or file should
be the genotype ID, columns 2 through \code{n.traits + 1} are trait columns, and subsequent columns are fixed effects.}

\item{n.traits}{The number of traits in the \code{pheno} data frame or file; each trait should be contained in a separate column.}

\item{delim}{The delimiter character in the pheno data file (if a filename is provided). For example,
"," is for csv files, and "\t" is for tab-delimited files.}
}
\value{
A variable of class \code{PopVar.data}
}
\description{
Combines marker genotype data and phenotype data into a single object for downstream
analysis and processing.
}
\details{
Genotypes in the \code{pheno} file or data frame that are not genotyped are removed from the analysis.
}
\examples{
# Read in phased cranberry marker data
geno_in <- read_geno(geno = cranberry_geno_phased, phased = TRUE, sep = "_hap", min.mac = 5)

# Create the PopVar.data object
popvar_data <- read_PopVar(geno = geno_in, pheno = cranberry_pheno, n.traits = 6)


}
